# spec/swagger.yaml
definitions:
...
  bots.Bot:
    properties:
      allow_guest:
        type: boolean
      avatar_url:
        type: string
      check_issue_count:
        type: integer
      check_state:
        type: string
      created_at:
        type: string
      display_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      metadata:
        additionalProperties: {}
        type: object
      owner_user_id:
        type: string
      status:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  bots.UpdateBotRequest:
    properties:
      avatar_url:
        type: string
      display_name:
        type: string
      is_active:
        type: boolean
      metadata:
        additionalProperties: {}
        type: object
    type: object
...
  handlers.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  message.Message:
    properties:
      bot_id:
        type: string
      content:
        items:
          type: integer
        type: array
      created_at:
        type: string
      external_message_id:
        type: string
      id:
        type: string
      metadata:
        additionalProperties: {}
        type: object
      platform:
        type: string
      role:
        type: string
      route_id:
        type: string
      sender_avatar_url:
        type: string
      sender_channel_identity_id:
        type: string
      sender_display_name:
        type: string
      sender_user_id:
        type: string
      source_reply_to_message_id:
        type: string
    type: object
...
info:
  contact: {}
  title: Memoh API
  version: 1.0.0
paths:
...
  /bots/{bot_id}/messages:
    get:
      description: List messages for a bot history with optional pagination
      parameters:
      - description: Bot ID
        in: path
        name: bot_id
        required: true
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Before
        in: query
        name: before
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/message.Message'
              type: array
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List bot history messages
      tags:
      - messages
...
  /bots/{id}:
    delete:
      description: Delete a bot user (owner/admin only)
      parameters:
      - description: Bot ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Delete bot
      tags:
      - bots
    get:
      description: Get a bot by ID (owner/admin only)
      parameters:
      - description: Bot ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bots.Bot'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get bot details
      tags:
      - bots
    put:
      description: Update bot profile (owner/admin only)
      parameters:
      - description: Bot ID
        in: path
        name: id
        required: true
        type: string
      - description: Bot update payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/bots.UpdateBotRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/bots.Bot'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Update bot details
      tags:
      - bots
...
swagger: "2.0"